TEST: splits config-form.spec.ts into three separate test files
